v1.2.0
- Refactor code base with new gcc-min
- Correct module variants' pointer
- Update tests to check all variants
- Update travis config
- Switch to codecov
