+++ .gitignore
@@ -11,3 +11,7 @@ logs
 node_modules
 coverage
 .nyc_output
+
+yarn.lock
+coverage.lcov
+package-lock.json
