+++ .travis.yml
@@ -1,9 +1,5 @@
 language: node_js
 node_js:
-  - "7.10"
-  - "7.6"
+  - "8.0"
   - "7.0"
-  - "6.5"
   - "6.0"
-after_success:
-  - npm run coveralls
