+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.1.0",
+  "version": "1.2.0",
   "name": "txtgen",
   "description": "Util for generating random sentences, paragraphs and articles in English",
   "homepage": "http://ndaidong.github.io/txtgen",
@@ -8,37 +8,58 @@
     "url": "https://github.com/ndaidong/txtgen"
   },
   "author": "@ndaidong",
-  "main": "./index.js",
+  "main": "./dist/bella.js",
+  "module": "./src/main",
+  "jsnext:main": "./src/main",
+  "browser": "./dist/txtgen.min.js",
   "engines": {
-    "node": ">= 6.0.0"
+    "node": ">= 6.0"
   },
   "scripts": {
-    "lint": "eslint .",
+    "lint": "eslint src tests",
     "nsp": "nsp check --output checkstyle",
-    "tape": "nyc tape test/start.js | tap-spec",
-    "report": "nyc report --reporter=lcov",
+    "tape": "nyc babel-node node_modules/.bin/tape tests/start.js | tap-spec",
+    "report": "nyc report --reporter=lcov > coverage.lcov && codecov",
     "pretest": "npm run lint && npm run nsp",
     "test": "npm run tape",
-    "posttest": "npm run report",
-    "coveralls": "npm test && cat ./coverage/lcov.info | coveralls",
-    "build": "gccmin"
+    "posttest": "CODECOV_TOKEN=0f80ffa2-ed0e-4501-ac91-49397be8b3fe npm run report",
+    "start": "babel-node src/main",
+    "build": "gccmin -e src/main.js -n txtgen -o dist -p package.json",
+    "reset": "node reset"
   },
   "devDependencies": {
-    "bellajs": "6.x.x",
-    "coveralls": "2.x.x",
+    "babel-plugin-transform-remove-strict-mode": "0.x.x",
+    "babel-preset-env": "1.x.x",
+    "codecov": "2.x.x",
     "eslint": "3.x.x",
     "eslint-config-ggc": "1.x.x",
-    "gcc-min": "3.x.x",
+    "gcc-min": "5.x.x",
+    "is": "3.x.x",
+    "jsdom": "11.x.x",
+    "jsdom-global": "3.x.x",
     "nsp": "2.x.x",
-    "nyc": "10.x.x",
+    "nyc": "11.x.x",
     "tap-spec": "4.x.x",
     "tape": "4.x.x"
   },
-  "gccmin": {
-    "source": "src",
-    "target": "dist",
-    "filename": "txtgen",
-    "globalVar": "txtgen"
+  "babel": {
+    "presets": [
+      [
+        "env",
+        {
+          "targets": {
+            "browsers": [
+              "> 1%"
+            ]
+          }
+        }
+      ]
+    ],
+    "plugins": [
+      [
+        "transform-remove-strict-mode"
+      ]
+    ]
   },
   "keywords": [
     "text",
