+++ tests/libs/jasmine-2.4.1/console.js
