+++ tests/libs/jasmine-2.4.1/jasmine_favicon.png
